:bug: 분리수거 정보 페이지가 subcategory를 인식하지 못하는 문제 수정
fetch 함수가 요청하는 api에 subCategory segment를 반영하지 않은 점 수정
